modif asso

+++ webservices/saveImage.php
@@ -5,11 +5,12 @@ function saveImage($fichier,$dossier){
 	if($ext==='jpeg') $ext='jpg';
 	$dest_fichier = $fichier.'.'.$ext;
 	$dest_dossier=$GLOBALS['root']."/".$dossier;
-	move_uploaded_file($_FILES[$dossier]['tmp_name'], $dest_dossier."/".$dest_fichier);
-	if(!($ext==='jpg')){
-		$img = new ImageConverter($dest_dossier ."/". $dest_fichier,'jpg');
-		rename($fichier.'.jpg',$dest_dossier ."/".$fichier.'.jpg');
-		unlink($dest_dossier ."/". $dest_fichier);
+	if (move_uploaded_file($_FILES[$dossier]['tmp_name'], $dest_dossier."/".$dest_fichier)){
+		if(!($ext==='jpg')){
+			$img = new ImageConverter($dest_dossier ."/". $dest_fichier,'jpg');
+			rename($fichier.'.jpg',$dest_dossier ."/".$fichier.'.jpg');
+			unlink($dest_dossier ."/". $dest_fichier);
+		}
 	}
 }
